Move thumbnail related components into separate file
`Video.tsx` got quite crowded. This commit only moves things around.
